Fix SVG Height to max of column heights.
Currently, the height of the SVG is set based on the height of the first column. This works when the first round of the bracket is completely full. However, if the bracket is sparse (i.e. bye rounds are not shown), then it's possible for the first column to have fewer matches (and therefore be shorter) than later columns.

To fix this, I have updated the calls to `calculateSVGDimensions`. Before, the numOfRows param was always set to the length of columns[0]. Now, it is set to the maximum length of all columns.